requirements.txt
amqp==1.4.8
anyjson==0.3.3
asyncio==3.4.3
billiard==3.3.0.22
boto==2.39.0
celery==3.1.19
click==6.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-boto==0.3.9
django-celery==3.1.17
django-cors-headers==1.1.0
django-enumfield==1.2.1
django-haystack==2.4.1
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.2.4
djangorestframework-jsonp==1.0.2
drf-haystack==1.5.6
elasticsearch==2.3.0
elasticsearch-curator==3.4.1
gunicorn==19.3.0
kombu==3.0.32
Pillow==3.1.1
psycopg2==2.6.1
PyMySQL==0.7.2
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
redis==2.10.5
rq==0.5.6
six==1.10.0
SQLAlchemy==1.0.8
static3==0.6.1
tzlocal==1.2
urllib3==1.14
websockets==2.7
